% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steel.Dwass.test.R
\name{Steel.Dwass}
\alias{Steel.Dwass}
\title{Steel-Dwass pairwise ranking test}
\usage{
Steel.Dwass(x, group)
}
\arguments{
\item{x}{data, with more than 3 groups}

\item{group}{group data,dataset is similar with those when you run ANOVA etc.}
}
\description{
The Steel Dwass method performs the multiple comparisons whilst controlling the overall experiment-wise error rate (it is the non-parametric equivalent to the Tukey All-Pairs method)
adapted from http://aoki2.si.gunma-u.ac.jp/R/Steel-Dwass.html
}
\examples{
library(Steel.Dwass.test)
x <- c(6.9, 7.5, 8.5, 8.4, 8.1, 8.7, 8.9, 8.2, 7.8, 7.3, 6.8,9.6, 9.4, 9.5, 8.5, 9.4, 9.9, 8.7, 8.1, 7.8, 8.8,5.7, 6.4, 6.8, 7.8, 7.6, 7.0, 7.7, 7.5, 6.8, 5.9,7.6, 8.7, 8.5, 8.5, 9.0, 9.2, 9.3, 8.0, 7.2, 7.9, 7.8)
group <- rep(1:4, c(11, 10, 10, 11))
Steel.Dwass(x, group)
}
\author{
Robert George Douglas Steel,Shigenobu Aoki,Weiping Mei
}
